Tasks

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,32 +1,33 @@
 #!/usr/bin/python3
-
 """
-This module prints the first State object from the database hbtn_0e_6_usa
+This script deletes all State objects
+with a name containing the letter `a`
+from the database `hbtn_0e_6_usa`.
 """
 
 from sys import argv
-from model_state import Base, State
+from model_state import State, Base
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-if __name__ == '__main__':
+if __name__ == "__main__":
+    """
+    Deletes State objects on the database.
+    """
 
-    usr = argv[1]
-    pswd = argv[2]
-    db = argv[3]
+    db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
+        argv[1], argv[2], argv[3])
 
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
-        usr, pswd, db), pool_pre_ping=True)
-
-    Base.metadata.create_all(engine)
-
+    engine = create_engine(db_url)
     Session = sessionmaker(bind=engine)
+
     session = Session()
 
-    """Filter by searching for patterns"""
-    states = session.query(State).filter(State.name.like('%a%')).all()
-    for state in states:
-        session.delete(state)
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
 
     session.commit()
+
     session.close()
